Migrate to Next.js app directory structure
Refactored project from pages/ to app/ directory, updating API routes, page components, and layout to use Next.js app directory conventions. Added new error, loading, and not-found components, updated metadata handling, and removed deprecated pages. Updated next.config.js for new dependencies and image qualities, and improved code structure for maintainability.
