+++ src/app/api/hello/route.ts
@@ -1,0 +1,10 @@
+import { NextResponse } from 'next/server'
+
+export async function GET() {
+  const data = {
+    name: 'Ahnaf An Nafee',
+    status: 'Alive'
+  }
+
+  return NextResponse.json(data)
+}
