+++ src/app/api/og/route.tsx
@@ -1,12 +1,10 @@
 /* eslint-disable @next/next/no-img-element */
-import { ImageResponse } from '@vercel/og'
+import { ImageResponse } from 'next/og'
 import { NextRequest } from 'next/server'
 
-export const config = {
-  runtime: 'edge'
-}
+export const runtime = 'edge'
 
-export default async function handler(req: NextRequest) {
+export async function GET(req: NextRequest) {
   try {
     const { searchParams } = new URL(req.url)
 
@@ -15,35 +13,7 @@ export default async function handler(req: NextRequest) {
     const subtitle = searchParams.get('subtitle') || 'PhD Student in AI & 3D Graphics @ GMU | Building Immersive Worlds'
     const theme = searchParams.get('theme') || 'default'
 
-    // Theming palette: default (purple) vs emerald variant to match homepage vibe
-    const palette =
-      theme === 'emerald' || theme === 'home'
-        ? {
-            rootBgImage: 'linear-gradient(135deg, #041E1A 0%, #064e3b 50%, #041E1A 100%)',
-            rootBgColor: '#041E1A',
-            panelBg: '#052e2b',
-            panelBorder: '#0f766e',
-            headerBg: '#064e3b',
-            headerBorder: '#0f766e',
-            accent: '#10b981',
-            iconGrad1: 'linear-gradient(135deg, #10b981, #34d399)',
-            iconGrad2: 'linear-gradient(135deg, #34d399, #6ee7b7)',
-            iconGrad3: 'linear-gradient(135deg, #059669, #10b981)'
-          }
-        : {
-            rootBgImage: 'linear-gradient(135deg, #0f172a 0%, #581c87 50%, #0f172a 100%)',
-            rootBgColor: '#0f172a',
-            panelBg: '#111827',
-            panelBorder: '#374151',
-            headerBg: '#1f2937',
-            headerBorder: '#374151',
-            accent: '#c084fc',
-            iconGrad1: 'linear-gradient(135deg, #3b82f6, #8b5cf6)',
-            iconGrad2: 'linear-gradient(135deg, #8b5cf6, #ec4899)',
-            iconGrad3: 'linear-gradient(135deg, #ec4899, #ef4444)'
-          }
-
-    const image = new ImageResponse(
+    return new ImageResponse(
       (
         <div
           style={{
@@ -53,8 +23,8 @@ export default async function handler(req: NextRequest) {
             flexDirection: 'column',
             alignItems: 'center',
             justifyContent: 'center',
-            backgroundColor: palette.rootBgColor,
-            backgroundImage: palette.rootBgImage,
+            backgroundColor: '#0f172a',
+            backgroundImage: 'linear-gradient(135deg, #0f172a 0%, #581c87 50%, #0f172a 100%)',
             padding: '40px'
           }}
         >
@@ -65,9 +35,9 @@ export default async function handler(req: NextRequest) {
               flexDirection: 'column',
               width: '100%',
               height: '100%',
-              backgroundColor: palette.panelBg,
+              backgroundColor: '#111827',
               borderRadius: '12px',
-              border: `1px solid ${palette.panelBorder}`,
+              border: '1px solid #374151',
               overflow: 'hidden',
               boxShadow: '0 25px 50px -12px rgba(0, 0, 0, 0.25)'
             }}
@@ -78,9 +48,9 @@ export default async function handler(req: NextRequest) {
                 display: 'flex',
                 alignItems: 'center',
                 justifyContent: 'space-between',
-                backgroundColor: palette.headerBg,
+                backgroundColor: '#1f2937',
                 padding: '12px 16px',
-                borderBottom: `1px solid ${palette.headerBorder}`
+                borderBottom: '1px solid #374151'
               }}
             >
               <div style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
@@ -117,7 +87,7 @@ export default async function handler(req: NextRequest) {
               {/* File Content */}
               <div style={{ display: 'flex', flexDirection: 'column', flex: 1, justifyContent: 'center' }}>
                 <div style={{ marginBottom: '24px' }}>
-                  <div style={{ color: palette.accent, fontSize: '24px', marginBottom: '8px' }}># {title}</div>
+                  <div style={{ color: '#c084fc', fontSize: '24px', marginBottom: '8px' }}># {title}</div>
                   <div style={{ color: '#d1d5db', fontSize: '18px', lineHeight: 1.6, marginBottom: '16px' }}>
                     {subtitle}
                   </div>
@@ -128,7 +98,7 @@ export default async function handler(req: NextRequest) {
                     <img
                       width='80'
                       height='80'
-                      style={{ borderRadius: '50%', border: `2px solid ${palette.accent}`, objectFit: 'cover' }}
+                      style={{ borderRadius: '50%', border: '2px solid #c084fc', objectFit: 'cover' }}
                       src='https://ik.imagekit.io/8ieg70pvks/profile?tr=w-160,h-160'
                       alt='Ahnaf An Nafee'
                     />
@@ -147,7 +117,7 @@ export default async function handler(req: NextRequest) {
                         style={{
                           width: '48px',
                           height: '48px',
-                          background: palette.iconGrad1,
+                          background: 'linear-gradient(135deg, #3b82f6, #8b5cf6)',
                           borderRadius: '8px',
                           display: 'flex',
                           alignItems: 'center',
@@ -161,7 +131,7 @@ export default async function handler(req: NextRequest) {
                         style={{
                           width: '48px',
                           height: '48px',
-                          background: palette.iconGrad2,
+                          background: 'linear-gradient(135deg, #8b5cf6, #ec4899)',
                           borderRadius: '8px',
                           display: 'flex',
                           alignItems: 'center',
@@ -175,7 +145,7 @@ export default async function handler(req: NextRequest) {
                         style={{
                           width: '48px',
                           height: '48px',
-                          background: palette.iconGrad3,
+                          background: 'linear-gradient(135deg, #ec4899, #ef4444)',
                           borderRadius: '8px',
                           display: 'flex',
                           alignItems: 'center',
@@ -186,7 +156,7 @@ export default async function handler(req: NextRequest) {
                         ⚡
                       </div>
                     </div>
-                    <div style={{ color: palette.accent, fontSize: '14px' }}>www.ahnafnafee.dev</div>
+                    <div style={{ color: '#c084fc', fontSize: '14px' }}>www.ahnafnafee.dev</div>
                   </div>
                 </div>
               </div>
@@ -208,15 +178,6 @@ export default async function handler(req: NextRequest) {
         height: 600
       }
     )
-    // Strongly prevent CDN/browser caching so style changes reflect immediately
-    image.headers.set(
-      'Cache-Control',
-      'no-store, no-cache, must-revalidate, proxy-revalidate, max-age=0, s-maxage=0'
-    )
-    image.headers.set('Pragma', 'no-cache')
-    image.headers.set('Expires', '0')
-    image.headers.set('Content-Type', 'image/png')
-    return image
   } catch (err) {
     console.info(JSON.stringify(err))
     return new Response('Failed to generate the og image', {
