+++ src/app/layout.tsx
@@ -1,0 +1,32 @@
+import { Analytics } from '@vercel/analytics/react'
+import { SpeedInsights } from '@vercel/speed-insights/next'
+import { ThemeProvider } from 'next-themes'
+import { Inter } from 'next/font/google'
+import { Toaster } from 'react-hot-toast'
+import { Header } from '@/components/UI/common'
+
+import '@/styles/globals.css'
+import 'prismjs/themes/prism.css'
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  preload: true,
+  variable: '--font-inter'
+})
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang='en' suppressHydrationWarning>
+      <body className={inter.className}>
+        <ThemeProvider attribute='class' defaultTheme='system' enableSystem>
+          <Header />
+          {children}
+          <Toaster position='bottom-center' />
+          <Analytics />
+          <SpeedInsights />
+        </ThemeProvider>
+      </body>
+    </html>
+  )
+}
