+++ src/app/loading.tsx
@@ -1,0 +1,7 @@
+export default function Loading() {
+  return (
+    <div className='flex items-center justify-center min-h-screen'>
+      <div className='animate-spin rounded-full h-32 w-32 border-b-2 border-primary-500'></div>
+    </div>
+  )
+}
